requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-camel-case==1.4.2
psycopg2==2.9.7
sqlparse==0.4.4
typing_extensions==4.7.1
